+++ src/types.ts
@@ -15,10 +15,8 @@ export type FilterOperator =
     | 'nbetween'
     | 'empty'
     | 'nempty'
-    | 'intersects'
-    | 'nintersects'
-    | 'intersects_bbox'
-    | 'nintersects_bbox';
+    | 'submitted'
+    | 'regex';
 
 export type ClientFilterOperator = FilterOperator | 'starts_with' | 'nstarts_with' | 'ends_with' | 'nends_with';
 
@@ -29,7 +27,7 @@ export type LogicalFilterAND = { _and: Filter[] };
 export type LogicalFilter = LogicalFilterOR | LogicalFilterAND;
 
 export type FieldFilter = {
-    [field: string]: FieldFilterOperator | FieldValidationOperator | FieldFilter;
+    [field: string]: FieldFilterOperator | FieldFilter;
 };
 
 export type FilterContext = Record<string, any>
@@ -45,19 +43,16 @@ export type FieldFilterOperator = {
     _nin?: (string | number)[];
     _null?: boolean;
     _nnull?: boolean;
+    _starts_with?: string;
+    _nstarts_with?: string;
+    _ends_with?: string;
+    _nends_with?: string;
     _contains?: string;
     _ncontains?: string;
     _between?: (string | number)[];
     _nbetween?: (string | number)[];
     _empty?: boolean;
     _nempty?: boolean;
-    _intersects?: string;
-    _nintersects?: string;
-    _intersects_bbox?: string;
-    _nintersects_bbox?: string;
-};
-
-export type FieldValidationOperator = {
     _submitted?: boolean;
     _regex?: string;
 };